feat: create proxy struct
